+++ source/extensions/isaacsim.examples.interactive/isaacsim/examples/interactive/tests/test_simple_stack.py
@@ -10,6 +10,8 @@
 
 import asyncio
 
+import carb
+import numpy as np
 import omni.kit
 
 # NOTE:
@@ -34,6 +36,19 @@ class TestSimpleStackExampleExtension(omni.kit.test.AsyncTestCase):
         await update_stage_async()
         while is_stage_loading():
             await update_stage_async()
+
+        settings = carb.settings.get_settings()
+        settings.set("/app/player/useFixedTimeStepping", False)
+        settings.set("/app/runLoops/main/rateLimitEnabled", False)
+
+        try:
+            import omni.kit.loop._loop as omni_loop
+
+            self._loop_runner = omni_loop.acquire_loop_interface()
+            self._loop_runner.set_manual_mode(True)
+        except Exception:
+            pass
+
         return
 
     # After running each test
@@ -54,8 +69,14 @@ class TestSimpleStackExampleExtension(omni.kit.test.AsyncTestCase):
         await self._sample._on_stacking_event_async()
         await update_stage_async()
         # run for 2500 frames and print time
-        for i in range(1000):
+        for i in range(1750):
             await update_stage_async()
+        cube_names = self._sample._franka_task.get_cube_names()
+        task_observations = self._sample._franka_task.get_observations()
+        for cube_name in cube_names:
+            self.assertTrue(
+                np.isclose(task_observations[cube_name]["position"][0:2], np.array([0.5, -0.5]), atol=1e2).all()
+            )
         pass
 
     async def test_reset(self):